+++ lib/cocoapods/installer/xcode/pods_project_generator/aggregate_target_installer.rb
@@ -119,7 +119,7 @@ module Pod
           #
           def create_copy_resources_script
             path = target.copy_resources_script_path
-            generator = Generator::CopyResourcesScript.new(target.resources_by_config, target.platform)
+            generator = Generator::CopyResourcesScript.new(target.resource_paths_by_config, target.platform)
             generator.save_as(path)
             add_file_to_support_group(path)
           end